feat(refresh interval): update rendering of HomeScreen
- refactored Carousel, Service and About sections to single components
- added new refresh time logics to use for smarter fetch policy in all of them
  - every component, screen and content needs their own unique
    refresh time key, that will be stored in AsyncStorage and used
    to individually take care of refresh times
  - as our queries are also unique per component/screen and content
    we can build a refresh time key based on the query params
