+++ src/components/screens/index.js
@@ -1,3 +1,5 @@
+export * from './About';
+export * from './Carousel';
 export * from './EventRecord';
 export * from './InfoCard';
 export * from './NewsItem';
@@ -5,4 +7,5 @@ export * from './PointOfInterest';
 export * from './PriceCard';
 export * from './OpeningTimesCard';
 export * from './OperatingCompanyInfo';
+export * from './Service';
 export * from './Tour';
